document_chat_bot/requirement.txt
openai
langchain==0.1.6 
langchain-community==0.0.19 
langchain-core==0.1.23
python-docx
tiktoken
faiss-cpu
tabula-py
docx2txt